requirements.txt
aenum==3.1.16
aiobotocore==2.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioitertools==0.12.0
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
apprise==1.9.3
asgi-lifespan==2.1.0
asttokens==3.0.0
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
botocore==1.37.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
coolname==2.2.0
cryptography==44.0.3
dateparser==1.2.1
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
docker==7.1.0
-e git+https://github.com/Thanaraklee/dsi321_2025.git@5ea3ae3b14f6eda8f175906b92071a0aac705bec#egg=dsi321_2025
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.20.3
greenlet==3.2.1
griffe==1.7.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lakefs==0.10.0
lakefs-client==1.44.0
lakefs-sdk==1.54.0
lakefs-spec==0.11.3
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.4.3
narwhals==1.38.2
nest-asyncio==1.6.0
numpy==2.2.5
oauthlib==3.2.2
opentelemetry-api==1.32.1
orjson==3.10.18
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
parso==0.8.4
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
playwright==1.52.0
prefect==3.4.1
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.4
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
python-socks==2.7.1
python-stdnum==1.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.0.0
rpds-py==0.24.0
ruamel.yaml==0.18.10
s3fs==2025.3.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.2
streamlit==1.45.0
tenacity==9.1.2
text-unidecode==1.3
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.0.7
uuid7==0.1.0
uv==0.7.3
uvicorn==0.34.2
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
whenever==0.8.0
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
